requirements.txt
glfw
pyopengl
pyrr
Pillow
pygame
